Se realiza modificacion en el cierre para que acepte decimales

+++ lluviaBackEnd/lluviaBackEnd/Models/Cierre.cs
@@ -22,7 +22,7 @@ namespace lluviaBackEnd.Models
         public int idAlmacen { get; set; }
         public float efectivoDisponible { get; set; }
         public float montoCierre { get; set; }
-        public int ProductosDevueltos { get; set; }
+        public float ProductosDevueltos { get; set; }
         public float MontoTotalDevoluciones { get; set; }
     }
 }
